Added InfectedHumanEntity
InfectedHumanEntity's soul purpose is to infect players. When a player dies to their infection (as opposed to internal organ damage, or other sources) this entity spawns

Currently the entity has the skin of those who see it. I wanted it to have the skin of the player who died, but I'm not quite at that level of figuring that out yet. Maybe in a later commit? Unlikely, but if I have more time after I design everything else, I'll try.

Oh yea also zombies have a low chance to infect you that is entirely blocked off if you wear a full armor set of gold, or better.